% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_documentation.R
\docType{data}
\name{small_fnac_corpus}
\alias{small_fnac_corpus}
\title{Financial news in Chinese}
\format{A \code{tbl_df} with one row per article.}
\usage{
small_fnac_corpus
}
\description{
Financial news for a couple of companies in Chinese webscraped using a search engine.
}
\seealso{
\link{speeches}
}
\keyword{datasets}

